requirements.txt
email-validator
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
SQLAlchemy-serializer
pytest
requests
Flask-RESTful
